requirements.txt
streamlit==1.41.1
openai==1.68.2
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
pypdf==5.1.0
python-dotenv==1.0.1
faiss-cpu==1.9.0
langchain-chroma